Removed command_executor
